refactor name-space functions
