+++ ivy/functional/frontends/jax/numpy/mathematical_functions.py
@@ -1,0 +1,228 @@
+# local
+import ivy
+from ivy.functional.frontends.jax.func_wrapper import (
+    to_ivy_arrays_and_back,
+)
+from ivy.functional.frontends.jax.numpy import promote_types_of_jax_inputs
+
+
+@to_ivy_arrays_and_back
+def absolute(x):
+    return ivy.abs(x)
+
+
+abs = absolute
+
+
+@to_ivy_arrays_and_back
+def add(x1, x2):
+    x1, x2 = promote_types_of_jax_inputs(x1, x2)
+    return ivy.add(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def arctan(x):
+    ret = ivy.atan(x)
+    return ret
+
+
+@to_ivy_arrays_and_back
+def arctan2(x1, x2):
+    return ivy.atan2(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def cos(x):
+    return ivy.cos(x)
+
+
+@to_ivy_arrays_and_back
+def cosh(x):
+    return ivy.cosh(x)
+
+
+@to_ivy_arrays_and_back
+def dot(a, b, *, precision=None):
+    a, b = promote_types_of_jax_inputs(a, b)
+    return ivy.matmul(a, b)
+
+
+@to_ivy_arrays_and_back
+def floor(x):
+    return ivy.floor(x)
+
+
+@to_ivy_arrays_and_back
+def mod(x1, x2):
+    return ivy.remainder(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def sinh(x):
+    return ivy.sinh(x)
+
+
+@to_ivy_arrays_and_back
+def sin(x):
+    return ivy.sin(x)
+
+
+@to_ivy_arrays_and_back
+def tan(x):
+    return ivy.tan(x)
+
+
+@to_ivy_arrays_and_back
+def tanh(x):
+    return ivy.tanh(x)
+
+
+@to_ivy_arrays_and_back
+def arccos(x):
+    return ivy.acos(x)
+
+
+@to_ivy_arrays_and_back
+def arccosh(x):
+    return ivy.acosh(x)
+
+
+@to_ivy_arrays_and_back
+def arcsin(x):
+    return ivy.asin(x)
+
+
+@to_ivy_arrays_and_back
+def arcsinh(x):
+    return ivy.asinh(x)
+
+
+@to_ivy_arrays_and_back
+def power(x1, x2):
+    x1, x2 = promote_types_of_jax_inputs(x1, x2)
+    return ivy.pow(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def trunc(x):
+    return ivy.trunc(x)
+
+
+@to_ivy_arrays_and_back
+def ceil(x):
+    return ivy.ceil(x)
+
+
+@to_ivy_arrays_and_back
+def float_power(x1, x2):
+    return ivy.float_power(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def deg2rad(x):
+    return ivy.deg2rad(x)
+
+
+@to_ivy_arrays_and_back
+def exp2(x):
+    return ivy.exp2(x)
+
+
+@to_ivy_arrays_and_back
+def gcd(x1, x2):
+    return ivy.gcd(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def i0(x):
+    return ivy.i0(x)
+
+
+@to_ivy_arrays_and_back
+def kron(a, b):
+    return ivy.kron(a, b)
+
+
+@to_ivy_arrays_and_back
+def lcm(x1, x2):
+    return ivy.lcm(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def logaddexp2(x1, x2):
+    return ivy.logaddexp2(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def trapz(y, x=None, dx=1.0, axis=-1, out=None):
+    return ivy.trapz(y, x=x, dx=dx, axis=axis, out=out)
+
+
+@to_ivy_arrays_and_back
+def sqrt(x, /):
+    return ivy.sqrt(x)
+
+
+@to_ivy_arrays_and_back
+def arctanh(x):
+    return ivy.atanh(x)
+
+
+@to_ivy_arrays_and_back
+def multiply(x1, x2):
+    return ivy.multiply(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def log10(x):
+    return ivy.log10(x)
+
+
+@to_ivy_arrays_and_back
+def logaddexp(x1, x2):
+    return ivy.logaddexp(x1, x2)
+
+
+@to_ivy_arrays_and_back
+def degrees(x):
+    return ivy.rad2deg(x)
+
+
+@to_ivy_arrays_and_back
+def negative(
+    x,
+    /,
+):
+    return ivy.negative(x)
+
+
+@to_ivy_arrays_and_back
+def rad2deg(
+    x,
+    /,
+):
+    return ivy.rad2deg(x)
+
+
+@to_ivy_arrays_and_back
+def tensordot(a, b, axes=2):
+    return ivy.tensordot(a, b, axes=axes)
+
+
+@to_ivy_arrays_and_back
+def divide(x1, x2, /):
+    if ivy.dtype(x1) == "int64" or ivy.dtype(x1) == "uint64":
+        x1 = ivy.astype(x1, ivy.float64)
+    x1, x2 = promote_types_of_jax_inputs(x1, x2)
+    return ivy.divide(x1, x2)
+
+
+true_divide = divide
+
+
+@to_ivy_arrays_and_back
+def exp(
+    x,
+    /,
+):
+    return ivy.exp(x)
